LSTM deployed
Fixed LSTM and move it back to src folder
