+++ data/semeval/semeval_label.pkl
@@ -1,0 +1,281 @@
+(dp0
+S' '
+p1
+I5
+sS'$'
+p2
+I65
+sS'('
+p3
+I70
+sS','
+p4
+I24
+sS'0'
+p5
+I62
+sS'4'
+p6
+I47
+sS'8'
+p7
+I73
+sS'@'
+p8
+I83
+sS'D'
+p9
+I28
+sS'H'
+p10
+I41
+sS'L'
+p11
+I48
+sS'P'
+p12
+I31
+sS'T'
+p13
+I44
+sS'X'
+p14
+I72
+sS'\\'
+p15
+I90
+sS'`'
+p16
+I80
+sS'd'
+p17
+I13
+sS'h'
+p18
+I1
+sS'l'
+p19
+I16
+sS'p'
+p20
+I29
+sS't'
+p21
+I17
+sS'x'
+p22
+I50
+sS'|'
+p23
+I81
+sS"'"
+p24
+I22
+sS'+'
+p25
+I59
+sS'/'
+p26
+I27
+sS'3'
+p27
+I61
+sS'7'
+p28
+I69
+sS';'
+p29
+I55
+sS'?'
+p30
+I32
+sS'C'
+p31
+I0
+sS'G'
+p32
+I46
+sS'K'
+p33
+I53
+sS'O'
+p34
+I42
+sS'S'
+p35
+I33
+sS'W'
+p36
+I60
+sS'['
+p37
+I75
+sS'_'
+p38
+I85
+sS'c'
+p39
+I20
+sS'g'
+p40
+I15
+sS'k'
+p41
+I21
+sS'o'
+p42
+I7
+sS's'
+p43
+I4
+sS'w'
+p44
+I8
+sS'\n'
+p45
+I91
+sS'"'
+p46
+I77
+sS'&'
+p47
+I54
+sS'*'
+p48
+I79
+sS'.'
+p49
+I34
+sS'2'
+p50
+I45
+sS'6'
+p51
+I68
+sS':'
+p52
+I63
+sS'B'
+p53
+I6
+sS'F'
+p54
+I36
+sS'J'
+p55
+I49
+sS'N'
+p56
+I51
+sS'R'
+p57
+I67
+sS'V'
+p58
+I78
+sS'Z'
+p59
+I74
+sS'^'
+p60
+I86
+sS'b'
+p61
+I19
+sS'f'
+p62
+I25
+sS'j'
+p63
+I58
+sS'n'
+p64
+I9
+sS'r'
+p65
+I2
+sS'v'
+p66
+I38
+sS'z'
+p67
+I30
+sS'~'
+p68
+I87
+sS'\t'
+p69
+I92
+sS'!'
+p70
+I37
+sS'%'
+p71
+I82
+sS')'
+p72
+I71
+sS'-'
+p73
+I39
+sS'1'
+p74
+I56
+sS'5'
+p75
+I57
+sS'9'
+p76
+I66
+sS'='
+p77
+I89
+sS'A'
+p78
+I26
+sS'E'
+p79
+I43
+sS'I'
+p80
+I40
+sS'M'
+p81
+I35
+sS'Q'
+p82
+I84
+sS'U'
+p83
+I52
+sS'Y'
+p84
+I23
+sS']'
+p85
+I76
+sS'a'
+p86
+I12
+sS'e'
+p87
+I11
+sS'i'
+p88
+I3
+sS'm'
+p89
+I10
+sS'q'
+p90
+I64
+sS'u'
+p91
+I18
+sS'y'
+p92
+I14
+sS'}'
+p93
+I88
+s.
