+++ data/senti_label.pkl
@@ -1,0 +1,281 @@
+(dp0
+S' '
+p1
+I0
+sS'$'
+p2
+I85
+sS'('
+p3
+I69
+sS','
+p4
+I36
+sS'0'
+p5
+I58
+sS'4'
+p6
+I67
+sS'8'
+p7
+I41
+sS'@'
+p8
+I76
+sS'D'
+p9
+I54
+sS'H'
+p10
+I38
+sS'L'
+p11
+I57
+sS'P'
+p12
+I60
+sS'T'
+p13
+I45
+sS'X'
+p14
+I81
+sS'\\'
+p15
+I89
+sS'`'
+p16
+I87
+sS'd'
+p17
+I22
+sS'h'
+p18
+I8
+sS'l'
+p19
+I3
+sS'p'
+p20
+I26
+sS't'
+p21
+I7
+sS'x'
+p22
+I33
+sS'|'
+p23
+I84
+sS"'"
+p24
+I27
+sS'+'
+p25
+I77
+sS'/'
+p26
+I55
+sS'3'
+p27
+I59
+sS'7'
+p28
+I46
+sS';'
+p29
+I30
+sS'?'
+p30
+I37
+sS'C'
+p31
+I31
+sS'G'
+p32
+I47
+sS'K'
+p33
+I49
+sS'O'
+p34
+I43
+sS'S'
+p35
+I32
+sS'W'
+p36
+I35
+sS'['
+p37
+I80
+sS'_'
+p38
+I75
+sS'c'
+p39
+I17
+sS'g'
+p40
+I24
+sS'k'
+p41
+I20
+sS'o'
+p42
+I14
+sS's'
+p43
+I13
+sS'w'
+p44
+I4
+sS'{'
+p45
+I92
+sS'\x7f'
+p46
+I90
+sS'&'
+p47
+I51
+sS'*'
+p48
+I79
+sS'.'
+p49
+I34
+sS'2'
+p50
+I66
+sS'6'
+p51
+I42
+sS':'
+p52
+I63
+sS'B'
+p53
+I29
+sS'F'
+p54
+I56
+sS'J'
+p55
+I64
+sS'N'
+p56
+I12
+sS'R'
+p57
+I44
+sS'V'
+p58
+I53
+sS'Z'
+p59
+I78
+sS'^'
+p60
+I86
+sS'b'
+p61
+I19
+sS'f'
+p62
+I23
+sS'j'
+p63
+I48
+sS'n'
+p64
+I15
+sS'r'
+p65
+I6
+sS'v'
+p66
+I28
+sS'z'
+p67
+I61
+sS'~'
+p68
+I72
+sS'\t'
+p69
+I88
+sS'!'
+p70
+I25
+sS'%'
+p71
+I83
+sS')'
+p72
+I70
+sS'-'
+p73
+I39
+sS'1'
+p74
+I40
+sS'5'
+p75
+I65
+sS'9'
+p76
+I68
+sS'='
+p77
+I71
+sS'A'
+p78
+I10
+sS'E'
+p79
+I62
+sS'I'
+p80
+I11
+sS'M'
+p81
+I9
+sS'Q'
+p82
+I73
+sS'U'
+p83
+I74
+sS'Y'
+p84
+I52
+sS']'
+p85
+I82
+sS'a'
+p86
+I2
+sS'e'
+p87
+I5
+sS'i'
+p88
+I16
+sS'm'
+p89
+I18
+sS'q'
+p90
+I50
+sS'u'
+p91
+I21
+sS'y'
+p92
+I1
+sS'}'
+p93
+I91
+s.
