--- testbed/run_lstm_char.sh
@@ -1,1 +1,0 @@
-THEANO_FLAGS=cuda.root=/usr/local/cuda,gcc.cxxflags='-march=x86-64',device=gpu,floatX=float32 python ./lstm_chars.py
